SPARKC-202: Switched Actor Streams with QueuedRDD Streams
The ActorStreamingSpec is removed and replaced with a non-actor test. We
have had a lot of difficulty getting these tests to consistently pass
due to actor failures and interactions with other parts of the system.
The queued RDD Dstream gives us direct control over what should be
happening, reduces overhead and allows the tests to execute much faster.
