+++ spark-cassandra-connector/src/test/scala/com/datastax/spark/connector/testkit/package.scala
@@ -12,10 +12,10 @@ package object testkit {
 
 
   val dataSeq = Seq (
-    Seq("1first ", "1round ", "1words "),
-    Seq("2second ", "2round ", "2words "),
-    Seq("3third ", "3round ", "3words "),
-    Seq("4fourth ", "4round ", "4words ")
+    Seq("1first", "1round", "1words"),
+    Seq("2second", "2round", "2words"),
+    Seq("3third", "3round", "3words"),
+    Seq("4fourth", "4round", "4words")
   )
 
   val data = dataSeq.head
